pyproject.toml
[tool.poetry]
name = "autodocgen"
version = "0.1.0"
description = "Automatically generate documentation for Python code using OpenAI models."
authors = ["Rohan Sobha <rohanraysobha@gmail.com>"]
readme = "README.md"
packages = [{include = "src/autodocgen"}]

[tool.poetry.dependencies]
python = "^3.9"
astunparse = "^1.6.3"
openai = "^0.27.7"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
astor = "^0.8.1"
black = "^23.3.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
